Fix remaining tests and update pyproject.toml.
