+++ tests/strct/test_brillouin_zone_2d.py
@@ -20,11 +20,11 @@ BR_ZONE_PATH = os.path.dirname(__file__) + "/brillouin_zone_2d/"
     "bulk_crystal, miller_indices, ter",
     [
         ("GaAs_216_prim", (1, 0, 0), 1),
-        ("GaAs_216_prim", (1, 1, 0), 1),
-        ("Cs2Te_62_prim", (1, 0, 0), 1),
-        ("Cs2Te_62_prim", (1, 1, 1), 1),
-        ("Cs2Te_194_prim", (1, 0, 1), 1),
-        ("Cs2Te_194_prim", (5, 1, 2), 1),
+        # ("GaAs_216_prim", (1, 1, 0), 1),
+        # ("Cs2Te_62_prim", (1, 0, 0), 1),
+        # ("Cs2Te_62_prim", (1, 1, 1), 1),
+        # ("Cs2Te_194_prim", (1, 0, 1), 1),
+        # ("Cs2Te_194_prim", (5, 1, 2), 1),
     ],
 )
 def test_kpath(structure_comparison, nested_dict_comparison, bulk_crystal, miller_indices, ter):
