+++ tests/strct/test_manipulation.py
@@ -8,11 +8,46 @@ import numpy as np
 import pytest
 
 # Internal library imports
-from aim2dat.strct import Structure
+from aim2dat.strct import Structure, StructureCollection, StructureOperations
+from aim2dat.io.yaml import load_yaml_file
 
 
 STRUCTURES_PATH = os.path.dirname(__file__) + "/structures/"
 STRUCTURE_MANIPULATION_PATH = os.path.dirname(__file__) + "/structure_manipulation/"
+
+
+@pytest.mark.parametrize("structure", ["Benzene"])
+def test_delete_atoms(structure_comparison, structure):
+    """Test delete atoms method."""
+    strct = Structure(
+        **dict(load_yaml_file(STRUCTURES_PATH + structure + ".yaml")), label="Benzene"
+    )
+    ref_p = load_yaml_file(STRUCTURE_MANIPULATION_PATH + structure + "_ref.yaml")
+    ref_p["structure"]["label"] = structure
+    new_strct = strct.delete_atoms(**ref_p["function_args"], change_label=True)
+    ref_p["structure"]["label"] += "_del"
+    structure_comparison(new_strct, ref_p["structure"])
+
+
+@pytest.mark.parametrize("structure", ["Cs2Te_62_prim", "GaAs_216_prim", "Cs2Te_19_prim_kinds"])
+def test_element_substitution(structure_comparison, structure):
+    """Test element substitution method."""
+    inputs = dict(load_yaml_file(STRUCTURES_PATH + structure + ".yaml"))
+    inputs["label"] = structure
+    inputs2 = dict(load_yaml_file(STRUCTURES_PATH + "Al_225_conv.yaml"))
+    inputs2["label"] = "Al_test"
+    ref_p = load_yaml_file(STRUCTURE_MANIPULATION_PATH + structure + "_ref.yaml")
+    strct_collect = StructureCollection()
+    strct_collect.append(**inputs)
+    strct_collect.append(**inputs2)
+    strct_ops = StructureOperations(strct_collect)
+    elements = ref_p["function_args"]["elements"]
+    elements = elements if isinstance(elements[0], (list, tuple)) else [elements]
+    subst_strct = strct_ops[0].substitute_elements(**ref_p["function_args"], change_label=True)
+    assert len(strct_ops.structures) == 2
+    structure_comparison(strct_ops.structures[0], inputs)
+    structure_comparison(strct_ops.structures[1], inputs2)
+    structure_comparison(subst_strct, ref_p["structure"])
 
 
 def test_scale_unit_cell_errors():
@@ -38,6 +73,21 @@ def test_scale_unit_cell_errors():
     assert (
         str(error.value) == "Provide either `scaling_factors` or `pressure` (with `bulk_modulus`)."
     )
+
+
+@pytest.mark.parametrize("new_label", ["GaAs_216_prim", "GaAs_216_prim_scaled-0.7"])
+def test_scale_unit_cell(structure_comparison, new_label):
+    """Test scale unit cell function."""
+    inputs = dict(load_yaml_file(STRUCTURES_PATH + "GaAs_216_prim.yaml"))
+    ref = dict(
+        load_yaml_file(STRUCTURE_MANIPULATION_PATH + "GaAs_216_prim_scale_unit_cell_ref.yaml")
+    )
+    ref["structure"]["label"] = new_label
+    strct = Structure(**inputs, label="GaAs_216_prim")
+    scaled_strct = strct.scale_unit_cell(
+        **ref["function_args"], change_label="scaled" in new_label
+    )
+    structure_comparison(scaled_strct, ref["structure"])
 
 
 def test_scale_unit_cell_uniform_scaling():
